+++ stac-cli/src/commands/item.rs
@@ -8,12 +8,14 @@ pub fn item(
     roles: Vec<String>,
     allow_relative_paths: bool,
     compact: bool,
+    disable_gdal: bool,
 ) -> Result<()> {
     let mut asset = Asset::new(href);
     asset.roles = roles;
     let item = Builder::new(id)
         .asset(key, asset)
         .canonicalize_paths(!allow_relative_paths)
+        .enable_gdal(!disable_gdal)
         .into_item()?;
     if compact {
         println!("{}", serde_json::to_string(&item)?);